Updated the map page to store integer values for start and end point
Added the functionality to send data to the controller from map page.
Provided methods to be called when data is recieved from the model via
controller.
Provided the method when delete is called on map page.
Displayed error message if user clicks on the get directions button
without selecting the start and end location.
In the application launch class, sending the controller reference to the
main view facade class.